+++ CHANGELOG.md
@@ -3,7 +3,8 @@
 ## [0.5.0]
 
 ### Added
-- True color to 256 color palette fallback
+- True color to 256 color palette fallback by using gchalk ansistyles package
+- Changed syntax of light colors (lightred ==> lred)
 
 
 ## [0.4.0] - 2026-04-06